@@ -53,40 +186,94 @@ set_target_properties(${GRAVITY_CGAME_LIBRARY} PROPERTIES COMPILE_FLAGS ${GRAVIT
 set_target_properties(${GRAVITY_CLIENT_EXECUTABLE} PROPERTIES COMPILE_FLAGS ${GRAVITY_CLIENT_FLAGS})
 set_target_properties(${GRAVITY_SERVER_EXECUTABLE} PROPERTIES COMPILE_FLAGS ${GRAVITY_SERVER_FLAGS})
 
-source_group("Header files" FILES ${GRAVITY_H})
-source_group("Source files" FILES ${GRAVITY_GAME_SRC})
-source_group("Source files" FILES ${GRAVITY_CGAME_SRC})
-source_group("Source files" FILES ${GRAVITY_CLIENT_SRC})
-source_group("Source files" FILES ${GRAVITY_SERVER_SRC})
+#source_group("Header files" FILES ${GRAVITY_H})
+#source_group("Source files" FILES ${GRAVITY_STUDIO_SRC})
+#source_group("Source files" FILES ${GRAVITY_GAME_SRC})
+#source_group("Source files" FILES ${GRAVITY_CGAME_SRC})
+#source_group("Source files" FILES ${GRAVITY_CLIENT_SRC})
+#source_group("Source files" FILES ${GRAVITY_SERVER_SRC})
+
+
+
+foreach(FILE
+#    ${GRAVITY_DEV_H}
+#    ${GRAVITY_DEV_SRC}
+#    ${GRAVITY_GAME_H}
+#    ${GRAVITY_GAME_SRC}
+#    ${GRAVITY_CGAME_H}
+#    ${GRAVITY_CGAME_SRC}
+#    ${GRAVITY_CLIENT_H}
+#    ${GRAVITY_CLIENT_SRC}
+#    ${GRAVITY_SERVER_H}
+#    ${GRAVITY_SERVER_SRC}
+#    ${GRAVITY_DATA_FILES}
+#    ${EXTLIBS_GAME_SRC}
+#    ${EXTLIBS_GAME_H}
+#    ${EXTLIBS_DEV_SRC}
+#    ${EXTLIBS_DEV_H}
+    )
+
+  get_filename_component(PARENT_DIR "${FILE}" PATH)
+
+   # skip src and include dir
+    string(REPLACE "${CMAKE_SOURCE_DIR}/include/" "" GROUP ${PARENT_DIR})
+    string(REPLACE "${CMAKE_SOURCE_DIR}/src/" "" GROUP ${GROUP})
+    string(REPLACE "${CMAKE_SOURCE_DIR}/" "" GROUP ${GROUP})
+
+  # skip src or include and changes /'s to \\'s
+  #string(REGEX REPLACE "(\\./)?(src|include)/?" "" GROUP "${PARENT_DIR}")
+   # string(REGEX REPLACE "
+  string(REPLACE "/" "\\" GROUP "${GROUP}")
+
+message("GROUP ${GROUP}")
+
+  # group into "Source Files" and "Header Files"
+  if ("${FILE}" MATCHES "(.*\\.cpp)|(.*\\.cc)|.*\\.c")
+    set(GROUP "Source Files\\${GROUP}")
+  elseif("${FILE}" MATCHES "(.*\\.h)|(.*\\.hpp)")
+    set(GROUP "Header Files\\${GROUP}")
+  endif()
+
+  source_group("${GROUP}" FILES "${FILE}")
+endforeach()
 
 if(MSVC)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /W3 /MP /fp:fast")
-    add_definitions(-D_CRT_SECURE_NO_WARNINGS)
+    #add_definitions(-D_CRT_SECURE_NO_WARNINGS)
 else()
-    set(CMAKE_CXX_FLAGS "-std=c++98 -Wall -Wextra -Wstrict-aliasing -pedantic")
+    set(CMAKE_CXX_FLAGS "-std=c++11 -Wall -Wextra -Wstrict-aliasing -pedantic -ffast-math")
 endif()
+
 add_definitions(-DGLM_FORCE_RADIANS=1)
+add_definitions(-DGRAVITY_COMPILETIME_SOURCE_PATH="${CMAKE_SOURCE_DIR}")
+add_definitions(-DBOOST_ALL_NO_LIB=1)
 
 find_package(Threads REQUIRED)
-find_package(Boost COMPONENTS program_options thread filesystem date_time)
+find_package(Boost COMPONENTS thread program_options system filesystem date_time)
 find_package(Box2D REQUIRED)
 find_package(OpenGL REQUIRED)
 find_package(GLEW REQUIRED)
+if (UNIX)
 set(SDL2_BUILDING_LIBRARY TRUE)
+endif()
 find_package(SDL2 REQUIRED)
 find_package(JsonCpp REQUIRED)
 find_package(PhysFS REQUIRED)
 find_package(GLM REQUIRED)
 #find_package(CEGUI REQUIRED)
+#find_package(LOGOG REQUIRED)
+find_package(YAMLCPP REQUIRED)
 
-message(${Boost_LIBRARIES})
-message(${BOX2D_LIBRARY})
-message(${OPENGL_LIBRARY})
-message(${GLEW_LIBRARY})
-message(${SDL2_LIBRARY})
-message(${JSONCPP_LIBRARY})
-message(${PHYSFS_LIBRARY})
-#message(${CEGUI_LIBRARY})
+#message(${Boost_LIBRARIES})
+#message(${BOX2D_LIBRARY})
+#message(${OPENGL_LIBRARY})
+#message(${GLEW_LIBRARY})
+#message(${SDL2_LIBRARY})
+#message(${JSONCPP_LIBRARY})
+#message(${PHYSFS_LIBRARY})
+##message(${CEGUI_LIBRARY})
+##message(${LOGOG_LIBRARY})
+#message(${YAMLCPP_LIBRARY})
 
 include_directories(SYSTEM ${Boost_INCLUDE_DIR})
 include_directories(SYSTEM ${BOX2D_INCLUDE_DIR})
