@@ -1,13 +1,17 @@
 /*
  * game/resource/resourceloader.cpp
  *
- * Copyright (c) 2014
+ * Copyright (c) 2014, 2015
  * Marcin Koziuk <marcin.koziuk@gmail.com>
  */
 
 #include <cstddef>
 #include <cstdlib>
+#include <vector>
+#include <istream>
+#include <fstream>
 #include <iostream>
+#include <memory>
 
 #include <boost/optional.hpp>
 #include <boost/none.hpp>
@@ -21,18 +25,42 @@
 #include <jsoncpp/json/reader.h>
 #endif
 
-#include "gravity/gravity.hpp"
-#include "gravity/logging.hpp"
-#include "gravity/game/resource/resourceloader.hpp"
+#include <gravity/gravity.hpp>
+#include <gravity/game/logging.hpp>
+#include <gravity/game/resource/resourceloader.hpp>
+
+#include <iterator>
+#include <algorithm>
 
 namespace Gravity {
-namespace Game {
 
 using boost::optional;
 using boost::none;
 using std::string;
 using std::vector;
 using std::size_t;
+
+static std::string FixPath(const std::string& path)
+{
+    std::string s = path;
+    std::replace(s.begin(), s.end(), '\\', '/');
+    return s;
+}
+
+ResourceLoader::StreamWrapper::StreamWrapper(PHYSFS_File *file)
+    : file(file)
+    , stream(file)
+{}
+
+ResourceLoader::StreamWrapper::~StreamWrapper()
+{
+    PHYSFS_close(file);
+}
+
+PhysFS::ifstream& ResourceLoader::StreamWrapper::operator*()
+{
+    return stream;
+}
 
 void ResourceLoader::Init()
 {
@@ -41,16 +69,42 @@ void ResourceLoader::Init()
     bool success = PHYSFS_mount("../gravity-ng/data", "/", true);
     if (!success) {
         LOG(fatal) << "Mounting of data directory failed";
+        throw std::runtime_error("Failed to mount PhysFS directory");
     }
 
-    LOG(info) << "Resource loader initialization successful";
-
-    char** paths = PHYSFS_getSearchPath();
     LOG(trace) << "Currently mounted search paths are:";
-    for (char** p = &paths[0]; *p != NULL; p++) {
-        LOG(trace) << *p;
+    const std::vector<std::string> paths = PhysFS::getSearchPath();
+    for (const std::string& path : paths) {
+        LOG(trace) << path;
     }
-    PHYSFS_freeList(paths);
+}
+
+optional<ResourceLoader::StreamWrapperPtr> ResourceLoader::OpenAsStream(const std::string& path)
+{
+    PHYSFS_File* file = PHYSFS_openRead(FixPath(path).c_str());
+    if (file) {
+        return std::make_shared<ResourceLoader::StreamWrapper>(file);
+    } else {
+        LOG(error) << "Resource not found: " << path;
+        return boost::none;
+    }
+}
+
+optional<std::vector<char>> ResourceLoader::OpenAsBytes(const std::string& path)
+{
+    PHYSFS_File* file = PHYSFS_openRead(FixPath(path).c_str());
+    if (file) {
+        size_t fileLength = PHYSFS_fileLength(file);
+        vector<char> data(fileLength);
+
+        PHYSFS_read(file, &data[0], fileLength, 1);
+        PHYSFS_close(file);
+
+        return data;
+    } else {
+        LOG(error) << "Resource not found: " << path;
+        return none;
+    }
 }
 
 optional<Json::Value> ResourceLoader::LoadJson(const string& key)
@@ -60,7 +114,7 @@ optional<Json::Value> ResourceLoader::LoadJson(const string& key)
         vector<char> data = maybeData.get();
         Json::Value root;
         Json::Reader reader;
-        reader.parse(&data[0], &data[data.size()], root);
+        reader.parse(&data.front(), &data.back(), root);
         return root;
     } else {
         return none;
@@ -89,5 +143,4 @@ void ResourceLoader::Deinit()
     PHYSFS_deinit();
 }
 
-} // namespace Game
 } // namespace Gravity
