@@ -5,11 +5,9 @@
  * Marcin Koziuk <marcin.koziuk@gmail.com>
  */
 
-#include "gravity/logging.hpp"
+#include "gravity/game/logging.hpp"
 #include "gravity/game/resource/resource.hpp"
 
 namespace Gravity {
-namespace Game {
 
-} // namespace Game
 } // namespace Gravity
